+++ pkg/alertmanager/alertstore/local/store.go
@@ -14,7 +14,8 @@ import (
 )
 
 const (
-	Name = "local"
+	Name         = "local"
+	templatesDir = "templates"
 )
 
 var (
@@ -148,9 +149,42 @@ func (f *Store) reloadConfigs() (map[string]alertspb.AlertConfigDesc, error) {
 		// The file name must correspond to the user tenant ID
 		user := strings.TrimSuffix(info.Name(), ext)
 
+		// Load template files
+		userTemplateDir := filepath.Join(f.cfg.Path, user, templatesDir)
+		var templates []*alertspb.TemplateDesc
+
+		if _, e := os.Stat(userTemplateDir); e == nil {
+			err = filepath.Walk(userTemplateDir, func(templatePath string, info os.FileInfo, err error) error {
+				if err != nil {
+					return errors.Wrapf(err, "unable to walk file path at %s", templatePath)
+				}
+				// Ignore files that are directories
+				if info.IsDir() {
+					return nil
+				}
+				content, err := os.ReadFile(templatePath)
+				if err != nil {
+					return errors.Wrapf(err, "unable to read alertmanager templates %s", templatePath)
+				}
+
+				templates = append(templates, &alertspb.TemplateDesc{
+					Body:     string(content),
+					Filename: info.Name(),
+				})
+				return nil
+			})
+
+			if err != nil {
+				return errors.Wrapf(err, "unable to list alertmanager templates: %s", userTemplateDir)
+			}
+		} else if !os.IsNotExist(e) {
+			return errors.Wrapf(e, "unable to read alertmanager templates %s", path)
+		}
+
 		configs[user] = alertspb.AlertConfigDesc{
 			User:      user,
 			RawConfig: string(content),
+			Templates: templates,
 		}
 		return nil
 	})